+85825f 9D.tas (40:14.901) from Splod
https://discord.com/channels/1155074438463754333/1187845552189231234/1207118942691393566
9D FULL LEVEL Draft in 40:14.901 (142053)
Ok
Important stuff, S&Q is used at the beginning of final checkpoint due to the insane timeactive bullshit that happens when you're 37 minutes into a level.
It can be moved to reuni to save RTA but I don't got time for that so im posting the file here
Speaking of reuni, it has like a billion frames of saving, most notably, a key I get there for drift reasons and a bunch of waiting that can probably be shaved down. If you move the s&q to reuni it will save rta and save more time due to the ability to get another key that's faster.
There's also a really anoying stun because of wierd moon block cycle bs that can also be cut with a s&q, but that means that some stuff in recon will have to be resynced.
MAC has some pretty bad movement, which i might touch on when i can
#xx-10-true-final-goodbye has some timesave i think if you smuggle the dash later (change inventory triggers allow you to keep a dash into dashless sections, currently it uses it for a bit of height but that's it)
This file is a synced version of the whole level, not nescesarily that optimized
I expect this to be sub 40 like tmr
I might post some easy improvements in a bit if i have time
+85825f